+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 - Fix incorrect QTableWidget method calls from `remove_row()` to `removeRow()` (nens/rana#3873)
 - Also upgrade geopackages upon loading (nens/threedi-schematisation-editor#486)
 - Add discharge capacity setting to orifice (nens/rana#4064)
+- Fix bug that caused the wrong model to be removed (nens/rana#4127)
 
 
 2.4.9 (2026-03-13)